requirements.txt
bcrypt==3.2.2
configparser==5.2.0
fastapi==0.75.2
passlib==1.7.4
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.0
python-jose==3.3.0
python-multipart==0.0.5
uvicorn==0.17.6
pymongo==4.1.1
requests-aws4auth==1.1.2
opensearch-py==1.1.0